# packages/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Start email-verified registration",
        "tags": ["Auth"]
      }
    },
    "/auth/verify-email/{token}": {
      "post": {
        "operationId": "verifyEmail",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyEmailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Verify a pending registration email token",
        "tags": ["Auth"]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "headers": {
              "Set-Cookie": {
                "description": "HTTP-only finnn_session session cookie",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Create an authenticated session cookie",
        "tags": ["Auth"]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "finnn_session": []
          }
        ],
        "summary": "Invalidate the current session cookie",
        "tags": ["Auth"]
      }
    },
    "/auth/session": {
      "get": {
        "operationId": "getSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "finnn_session": []
          }
        ],
        "summary": "Read the current authenticated session",
        "tags": ["Auth"]
      }
    },
    "/auth/user": {
      "patch": {
        "operationId": "updateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "finnn_session": []
          }
        ],
        "summary": "Update current user settings",
        "tags": ["Auth"]
      }
    },
    "/workspaces/{workspaceId}/invites": {
      "post": {
        "operationId": "createWorkspaceInvite",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInviteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateInviteResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "finnn_session": []
          }
        ],
        "summary": "Create a workspace invite",
        "tags": ["Workspaces"]
      }
    },
    "/workspace-invites/{token}": {
      "get": {
        "operationId": "getWorkspaceInvite",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceInvitePreviewResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Preview a workspace invite",
        "tags": ["Workspace Invites"]
      }
    },
    "/workspace-invites/{token}/accept": {
      "post": {
        "operationId": "acceptWorkspaceInvite",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptInviteResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "finnn_session": []
          }
        ],
        "summary": "Accept a workspace invite",
        "tags": ["Workspace Invites"]
      }
    },
    "/health": {
      "get": {
        "operationId": "getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Read API health status",
        "tags": ["Health"]
      }
    }
  },
  "info": {
    "title": "Finnn API",
    "description": "Backend API for Finnn finance tracking",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Health",
      "description": ""
    },
    {
      "name": "Auth",
      "description": ""
    },
    {
      "name": "Workspace Invites",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "finnn_session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "finnn_session"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Finn User",
            "maxLength": 100,
            "minLength": 1
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "correct-horse-battery",
            "minLength": 6
          }
        },
        "required": ["name", "email", "password"]
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        },
        "required": ["success"]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "path": {
            "type": "string",
            "example": "/auth/register"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-25T12:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": ["statusCode", "message", "error", "path", "timestamp"]
      },
      "VerifyEmailResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "userId": {
            "type": "string",
            "example": "665f5d865ef5a20c0d2f1111"
          }
        },
        "required": ["success", "userId"]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "correct-horse-battery",
            "minLength": 1
          }
        },
        "required": ["email", "password"]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665f5d865ef5a20c0d2f1111"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "example": "Finn User"
          },
          "image": {
            "type": "string",
            "example": "avatar-01",
            "nullable": true
          }
        },
        "required": ["id", "email", "name"]
      },
      "AuthUserResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          }
        },
        "required": ["user"]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "authenticated": {
            "type": "boolean",
            "example": true
          },
          "user": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthUserDto"
              }
            ]
          }
        },
        "required": ["authenticated"]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Finn User",
            "maxLength": 100,
            "minLength": 1
          },
          "image": {
            "type": "string",
            "example": "avatar-01",
            "nullable": true
          }
        },
        "required": ["name"]
      },
      "CreateInviteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "member@example.com"
          },
          "expiresInDays": {
            "type": "number",
            "default": 7,
            "maximum": 30,
            "minimum": 1
          }
        },
        "required": ["email"]
      },
      "WorkspaceInviteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665f5d865ef5a20c0d2f1111"
          },
          "workspaceId": {
            "type": "string",
            "example": "665f5d865ef5a20c0d2f2222"
          },
          "email": {
            "type": "string",
            "example": "member@example.com"
          },
          "token": {
            "type": "string",
            "example": "b6f13a6bead..."
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-06-01T12:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": ["id", "workspaceId", "email", "token", "expiresAt"]
      },
      "CreateInviteResponseDto": {
        "type": "object",
        "properties": {
          "invite": {
            "$ref": "#/components/schemas/WorkspaceInviteDto"
          }
        },
        "required": ["invite"]
      },
      "WorkspaceInvitePreviewDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "member@example.com"
          },
          "workspaceName": {
            "type": "string",
            "example": "Personal budget"
          },
          "workspaceId": {
            "type": "string",
            "example": "665f5d865ef5a20c0d2f2222"
          },
          "expiresAt": {
            "type": "string",
            "example": "2026-06-01T12:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": ["email", "workspaceName", "workspaceId", "expiresAt"]
      },
      "WorkspaceInvitePreviewResponseDto": {
        "type": "object",
        "properties": {
          "invite": {
            "$ref": "#/components/schemas/WorkspaceInvitePreviewDto"
          }
        },
        "required": ["invite"]
      },
      "AcceptInviteResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "workspaceId": {
            "type": "string",
            "example": "665f5d865ef5a20c0d2f2222"
          }
        },
        "required": ["success", "workspaceId"]
      },
      "HealthResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ok"],
            "example": "ok"
          }
        },
        "required": ["status"]
      }
    }
  }
}